+++ examples/Kafka/Avro/src/Avro.csproj
@@ -19,8 +19,6 @@
         <PackageReference Include="Amazon.Lambda.Core" Version="2.5.0"/>
         <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4"/>
         <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="2.0.0" />
-        <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="2.0.0" />
-
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\..\..\..\libraries\src\AWS.Lambda.Powertools.Kafka.Avro\AWS.Lambda.Powertools.Kafka.Avro.csproj" />
