+++ examples/Kafka/Json/src/Json.csproj
@@ -16,7 +16,6 @@
         <PackageReference Include="Amazon.Lambda.Core" Version="2.5.0"/>
         <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4"/>
         <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="2.0.0" />
-        <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="2.0.0" />
     </ItemGroup>
     <ItemGroup>
         <ProjectReference Include="..\..\..\..\libraries\src\AWS.Lambda.Powertools.Kafka.Json\AWS.Lambda.Powertools.Kafka.Json.csproj" />
