+++ examples/Kafka/Protobuf/src/Protobuf.csproj
@@ -16,7 +16,6 @@
         <PackageReference Include="Amazon.Lambda.Core" Version="2.5.0" />
         <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4" />
         <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="2.0.0" />
-        <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="2.0.0" />
         <PackageReference Include="Grpc.Tools" Version="2.72.0">
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
