--- examples/Kafka/src/Avro/AvroKey.avsc
@@ -1,24 +1,0 @@
-{
-  "namespace": "AWS.Lambda.Powertools.Kafka.Tests",
-  "type": "record",
-  "name": "AvroKey",
-  "fields": [
-    {
-      "name": "id",
-      "type": "int"
-    },
-    {
-      "name": "color",
-      "type": {
-        "type": "enum",
-        "name": "Color",
-        "symbols": [
-          "UNKNOWN",
-          "GREEN",
-          "RED"
-        ],
-        "default": "UNKNOWN"
-      }
-    }
-  ]
-}
