--- examples/Kafka/src/Avro/AvroProduct.avsc
@@ -1,10 +1,0 @@
-{
-  "namespace": "AWS.Lambda.Powertools.Kafka.Tests",
-  "type": "record",
-  "name": "AvroProduct",
-  "fields": [
-    {"name": "id", "type": "int"},
-    {"name": "name", "type": "string"},
-    {"name": "price", "type": "double"}
-  ]
-}
