--- examples/Kafka/src/Avro/kafka-avro-event.json
@@ -1,51 +1,0 @@
-{
-  "eventSource": "aws:kafka",
-  "eventSourceArn": "arn:aws:kafka:us-east-1:0123456789019:cluster/SalesCluster/abcd1234-abcd-cafe-abab-9876543210ab-4",
-  "bootstrapServers": "b-2.demo-cluster-1.a1bcde.c1.kafka.us-east-1.amazonaws.com:9092,b-1.demo-cluster-1.a1bcde.c1.kafka.us-east-1.amazonaws.com:9092",
-  "records": {
-    "mytopic-0": [
-      {
-        "topic": "mytopic",
-        "partition": 0,
-        "offset": 15,
-        "timestamp": 1545084650987,
-        "timestampType": "CREATE_TIME",
-        "key": "NDI=",
-        "value": "0g8MTGFwdG9wUrgehes/j0A=",
-        "headers": [
-          {
-            "headerKey": [104, 101, 97, 100, 101, 114, 86, 97, 108, 117, 101]
-          }
-        ]
-      },
-      {
-        "topic": "mytopic",
-        "partition": 0,
-        "offset": 16,
-        "timestamp": 1545084650988,
-        "timestampType": "CREATE_TIME",
-        "key": "NDI=",
-        "value": "1A8UU21hcnRwaG9uZVK4HoXrv4JA",
-        "headers": [
-          {
-            "headerKey": [104, 101, 97, 100, 101, 114, 86, 97, 108, 117, 101]
-          }
-        ]
-      },
-      {
-        "topic": "mytopic",
-        "partition": 0,
-        "offset": 17,
-        "timestamp": 1545084650989,
-        "timestampType": "CREATE_TIME",
-        "key": null,
-        "value": "1g8USGVhZHBob25lc0jhehSuv2JA",
-        "headers": [
-          {
-            "headerKey": [104, 101, 97, 100, 101, 114, 86, 97, 108, 117, 101]
-          }
-        ]
-      }
-    ]
-  }
-}
