--- examples/Kafka/src/Kafka.csproj
@@ -1,74 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net8.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <GenerateRuntimeConfigurationFiles>true</GenerateRuntimeConfigurationFiles>
-        <AWSProjectType>Lambda</AWSProjectType>
-        <!-- This property makes the build directory similar to a publish directory and helps the AWS .NET Lambda Mock Test Tool find project dependencies. -->
-        <CopyLocalLockFileAssemblies>true</CopyLocalLockFileAssemblies>
-        <!-- Generate ready to run images during publishing to improve cold start time. -->
-        <PublishReadyToRun>true</PublishReadyToRun>
-    </PropertyGroup>
-    <ItemGroup>
-        <PackageReference Include="Amazon.Lambda.RuntimeSupport" Version="1.12.3" />
-        <PackageReference Include="Amazon.Lambda.Core" Version="2.5.0" />
-        <PackageReference Include="Amazon.Lambda.Serialization.SystemTextJson" Version="2.4.4" />
-        <PackageReference Include="AWS.Lambda.Powertools.Logging" Version="2.0.0" />
-        <PackageReference Include="AWS.Lambda.Powertools.Metrics" Version="2.0.0" />
-        <PackageReference Include="Grpc.Tools" Version="2.72.0">
-          <PrivateAssets>all</PrivateAssets>
-          <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-    </ItemGroup>
-    <ItemGroup>
-        <ProjectReference Include="..\..\..\libraries\src\AWS.Lambda.Powertools.Kafka.Avro\AWS.Lambda.Powertools.Kafka.Avro.csproj" />
-        <ProjectReference Include="..\..\..\libraries\src\AWS.Lambda.Powertools.Kafka.Json\AWS.Lambda.Powertools.Kafka.Json.csproj" />
-        <ProjectReference Include="..\..\..\libraries\src\AWS.Lambda.Powertools.Kafka.Protobuf\AWS.Lambda.Powertools.Kafka.Protobuf.csproj" />
-        
-    </ItemGroup>
-    
-<!--    For Avro-->
-    <Target Name="GenerateAvroClasses" BeforeTargets="CoreCompile">
-        <Exec Command="avrogen -s $(ProjectDir)Avro/AvroProduct.avsc $(ProjectDir)Avro/Generated"/>
-        <Exec Command="avrogen -s $(ProjectDir)Avro/AvroKey.avsc $(ProjectDir)Avro/Generated"/>
-        <ItemGroup>
-            <Compile Include="$(ProjectDir)Avro/Generated/**/*.cs"/>
-        </ItemGroup>
-    </Target>
-    
-<!--    For Protobuf -->
-    <ItemGroup>
-        <None Remove="Protobuf\Key.proto" />
-        <Protobuf Include="Protobuf\Key.proto">
-            <GrpcServices>Client</GrpcServices>
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-            <Generator>MSBuild:Compile</Generator>
-        </Protobuf>
-        <None Remove="Protobuf\Product.proto" />
-        <Protobuf Include="Protobuf\Product.proto" GrpcServices="Client">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-
-        </Protobuf>
-    </ItemGroup>
-    
-    <ItemGroup>
-        <None Remove="Avro\kafka-avro-event.json" />
-        <Content Include="Avro\kafka-avro-event.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <None Remove="Json\kafka-json-event.json" />
-        <Content Include="Json\kafka-json-event.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-        <None Remove="Protobuf\kafka-protobuf-event.json" />
-        <Content Include="Protobuf\kafka-protobuf-event.json">
-            <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-        </Content>
-    </ItemGroup>
-    
-    <ItemGroup>
-      <Folder Include="Avro\Generated\AWS\Lambda\Powertools\Kafka\Tests\" />
-    </ItemGroup>
-</Project>
