--- examples/Kafka/src/Protobuf/Key.proto
@@ -1,14 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "TestKafka";
-
-message ProtobufKey {
-  int32 id = 1;
-  Color color = 2;
-}
-
-enum Color {
-  UNKNOWN = 0;
-  GREEN = 1;
-  RED = 2;
-}
