--- examples/Kafka/src/Protobuf/Product.proto
@@ -1,9 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "TestKafka";
-
-message ProtobufProduct {
-  int32 id = 1;
-  string name = 2;
-  double price = 3;
-}
