+++ examples/examples.sln
@@ -111,10 +111,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AOT_Logging.Tests", "AOT\AO
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Kafka", "Kafka", "{71027B81-CA39-498C-9A50-ADDAFA2AC2F5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kafka", "Kafka\src\Kafka.csproj", "{D9FF982F-7CD7-4652-94BB-B387B7FC034F}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{0D5246BB-02B3-43EE-9EE3-E0E627170C9B}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Json", "Kafka\Json\src\Json.csproj", "{58EC305E-353A-4996-A541-3CF7FC0EDD80}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Protobuf", "Kafka\Protobuf\src\Protobuf.csproj", "{853F6FE9-1762-4BA3-BAF4-2FCD605B81CF}"
@@ -214,10 +210,6 @@ Global
 		{FC010A0E-64A9-4440-97FE-DEDA8CEE0BE5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FC010A0E-64A9-4440-97FE-DEDA8CEE0BE5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FC010A0E-64A9-4440-97FE-DEDA8CEE0BE5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D9FF982F-7CD7-4652-94BB-B387B7FC034F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D9FF982F-7CD7-4652-94BB-B387B7FC034F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D9FF982F-7CD7-4652-94BB-B387B7FC034F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D9FF982F-7CD7-4652-94BB-B387B7FC034F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{58EC305E-353A-4996-A541-3CF7FC0EDD80}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{58EC305E-353A-4996-A541-3CF7FC0EDD80}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{58EC305E-353A-4996-A541-3CF7FC0EDD80}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -277,8 +269,6 @@ Global
 		{343CF6B9-C006-43F8-924C-BF5BF5B6D051} = {FE1CAA26-87E9-4B71-800E-81D2997A7B53}
 		{FC02CF45-DE15-4413-958A-D86808B99146} = {FEE72EAB-494F-403B-A75A-825E713C3D43}
 		{FC010A0E-64A9-4440-97FE-DEDA8CEE0BE5} = {F3480212-EE7F-46FE-9ED5-24ACAB5B681D}
-		{0D5246BB-02B3-43EE-9EE3-E0E627170C9B} = {71027B81-CA39-498C-9A50-ADDAFA2AC2F5}
-		{D9FF982F-7CD7-4652-94BB-B387B7FC034F} = {0D5246BB-02B3-43EE-9EE3-E0E627170C9B}
 		{58EC305E-353A-4996-A541-3CF7FC0EDD80} = {71027B81-CA39-498C-9A50-ADDAFA2AC2F5}
 		{853F6FE9-1762-4BA3-BAF4-2FCD605B81CF} = {71027B81-CA39-498C-9A50-ADDAFA2AC2F5}
 		{B03F22B2-315C-429B-9CC0-C15BE94CBF77} = {71027B81-CA39-498C-9A50-ADDAFA2AC2F5}
